requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.35
python-dotenv==1.0.1

langchain==0.3.7
langchain-community==0.3.6
langchain-openai==0.2.6

# For typing
typing-extensions>=4.12.2
